+++ composer/algorithms/__init__.py
@@ -67,6 +67,7 @@ from composer.algorithms.hparams import ProgressiveResizingHparams as Progressiv
 from composer.algorithms.hparams import RandAugmentHparams as RandAugmentHparams
 from composer.algorithms.hparams import SAMHparams as SAMHparams
 from composer.algorithms.hparams import ScaleScheduleHparams as ScaleScheduleHparams
+from composer.algorithms.hparams import SelectiveBackpropHparams as SelectiveBackpropHparams
 from composer.algorithms.hparams import SqueezeExciteHparams as SqueezeExciteHparams
 from composer.algorithms.hparams import StochasticDepthHparams as StochasticDepthHparams
 from composer.algorithms.label_smoothing import LabelSmoothing as LabelSmoothing
@@ -79,7 +80,6 @@ from composer.algorithms.randaugment import RandAugment as RandAugment
 from composer.algorithms.sam import SAM as SAM
 from composer.algorithms.scale_schedule import ScaleSchedule as ScaleSchedule
 from composer.algorithms.selective_backprop import SelectiveBackprop as SelectiveBackprop
-from composer.algorithms.selective_backprop import SelectiveBackpropHparams as SelectiveBackpropHparams
 from composer.algorithms.seq_length_warmup import SeqLengthWarmup as SeqLengthWarmup
 from composer.algorithms.seq_length_warmup import SeqLengthWarmupHparams as SeqLengthWarmupHparams
 from composer.algorithms.squeeze_excite import SqueezeExcite as SqueezeExcite
